subdivide files out of device
