+++ meerk40t/grbl/tcpoutput.py
@@ -1,0 +1,81 @@
+import socket
+import threading
+import time
+
+
+class TCPOutput:
+    def __init__(self, context):
+        super().__init__()
+        self.service = context
+        self._stream = None
+
+        self.lock = threading.RLock()
+        self.buffer = bytearray()
+        self.thread = None
+
+    def connect(self):
+        try:
+            self._stream = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+            self._stream.connect((self.service.address, self.service.port))
+            self.service.signal("tcp;status", "connected")
+        except (ConnectionError, TimeoutError):
+            self.disconnect()
+
+    def disconnect(self):
+        self.service.signal("tcp;status", "disconnected")
+        self._stream.close()
+        self._stream = None
+
+    def write(self, data):
+        self.service.signal("tcp;write", data)
+        if isinstance(data, str):
+            data = bytes(data, "utf-8")
+        with self.lock:
+            self.buffer += data
+            self.service.signal("tcp;buffer", len(self.buffer))
+        self._start()
+
+    realtime_write = write
+
+    def _start(self):
+        if self.thread is None:
+            self.thread = self.service.threaded(
+                self._sending,
+                thread_name=f"sender-{self.service.port}",
+                result=self._stop,
+            )
+
+    def _stop(self, *args):
+        self.thread = None
+
+    def _sending(self):
+        tries = 0
+        while True:
+            try:
+                if len(self.buffer):
+                    if self._stream is None:
+                        self.connect()
+                        if self._stream is None:
+                            return
+                    with self.lock:
+                        sent = self._stream.send(self.buffer)
+                        del self.buffer[:sent]
+                        self.service.signal("tcp;buffer", len(self.buffer))
+                    tries = 0
+                else:
+                    tries += 1
+                    time.sleep(0.1)
+            except (ConnectionError, OSError):
+                tries += 1
+                self.disconnect()
+                time.sleep(0.05)
+            if tries >= 20:
+                with self.lock:
+                    if len(self.buffer) == 0:
+                        break
+
+    def __repr__(self):
+        return f"TCPOutput('{self.service.address}:{self.service.port}')"
+
+    def __len__(self):
+        return len(self.buffer)
